samples/python/semantic-kernel-multiturn/requirements.txt
pydantic
semantic-kernel
microsoft-agents-authentication-msal
microsoft-agents-hosting-core
microsoft-agents-hosting-aiohttp
microsoft-agents-activity